+++ valueclient/sync_conn.go
@@ -32,7 +32,7 @@ func NewSyncConn() *syncConn {
 	return t
 }
 
-func (t *syncConn) connect(ctx context.Context, dialer valuerpc.Dialer, clientId int64, sessionToken string, credential value.Value, sendingCap int64, respHandler responseHandler, errorHandler ErrorHandler) error {
+func (t *syncConn) connect(ctx context.Context, dialer valuerpc.Dialer, clientId int64, resumeToken string, credential value.Value, sendingCap int64, respHandler responseHandler, errorHandler ErrorHandler) error {
 
 	t.connecting.Lock()
 	defer t.connecting.Unlock()
@@ -41,7 +41,7 @@ func (t *syncConn) connect(ctx context.Context, dialer valuerpc.Dialer, clientId
 		return nil
 	}
 
-	conn, err := newConn(ctx, dialer, clientId, sessionToken, credential, sendingCap, respHandler, errorHandler)
+	conn, err := newConn(ctx, dialer, clientId, resumeToken, credential, sendingCap, respHandler, errorHandler)
 	if err != nil {
 		return err
 	}